moved ES_URL and REST_API endpoints to general config/index.js
map display will default to true, removed variable from config/tosca
