+++ src/config/index.template.js
@@ -1,0 +1,15 @@
+// GRQ's ES url
+exports.GRQ_ES_URL = "http://localhost:9200";
+
+// GRQ's Rest API
+exports.GRQ_API_BASE = "http://localhost:8878"; // base url for GRQ API
+exports.GRQ_REST_API_V1 = `${this.GRQ_API_BASE}/api/v0.1`;
+exports.GRQ_REST_API_V2 = `${this.GRQ_API_BASE}/api/v0.2`;
+
+// Mozart's ES url
+exports.MOZART_ES_URL = "http://localhost:9300";
+
+// Mozart's Rest API
+exports.MOZART_REST_API_BASE = "http://localhost:8888";
+exports.MOZART_REST_API_V1 = `${this.MOZART_REST_API_BASE}/api/v0.1`;
+exports.MOZART_REST_API_V2 = `${this.MOZART_REST_API_BASE}/api/v0.2`;
