+++ src/pages/FigaroRuleEditor/index.jsx
@@ -16,7 +16,7 @@ import { Border, SubmitStatusBar } from "../../components/miscellaneous";
 
 import HeaderBar from "../../components/HeaderBar";
 
-import { MOZART_REST_API_V1 } from "../../config/figaro";
+import { MOZART_REST_API_V1 } from "../../config";
 
 import {
   validateQuery,
